requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
Faker==27.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
pillow==10.4.0
PyJWT==2.9.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
qrcode==7.4.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
